Update 10 files
- /V0.06/levelLoader.h
- /V0.06/animations.c
- /V0.06/LemonData/Animations/Spring.txt
- /V0.06/LemonData/LevelData/Level0Data.txt
- /V0.06/LemonData/LevelData/Level1Data.txt
- /V0.06/animations.h
- /V0.06/data.h
- /V0.06/gameObjects.c
- /V0.06/gameObjects.h
- /V0.06/levelLoader.c

+++ V0.06/data.h
@@ -443,7 +443,7 @@ struct displayData
 	int pixelXOffset;
 	int pixelYOffset;
 
-	short currentAnimation;
+	int currentAnimation;
 	int animationTick;
 	short animationLoopCount;
